requirements.txt
git+https://github.com/Rapptz/discord.py
discord-ext-ui
python-dotenv
